fix(config): extract state-dir-name helpers to leaf module to break paths<->utils import cycle (#45765)
utils.ts imported isExplicitOpenClawHomeStateDir from paths.ts, which imports the config types barrel, forming a madge cycle. Moved the pure state-dir-name helpers to src/config/state-dir-names.ts (node:path only); paths.ts re-exports for compatibility; utils.ts imports the leaf module directly. No behavior change.
